Simplify code for building UCCSD variational form
The ansatz is now built more from "first principles" than the previous code, which
followed the IBM tutorial for LiH. I verified that the circuit produced by this code
is identical to the one produced before.

Advantage of this simplified code is that it is a better interface for generalizing
to support more molecules.